+++ DashAI/back/models/base_model.py
@@ -1,9 +1,9 @@
-from abc import abstractmethod
+from abc import ABCMeta, abstractmethod
 
 from DashAI.back.config_object import ConfigObject
 
 
-class BaseModel(ConfigObject):
+class BaseModel(ConfigObject, metaclass=ABCMeta):
     """
     Abstract class of all machine learning models.
     All models must extend this clas and implement save and load methods.
@@ -11,6 +11,10 @@ class BaseModel(ConfigObject):
 
     MODEL: str
     SCHEMA: dict
+
+    @property
+    def _compatible_tasks(self) -> list:
+        raise NotImplementedError
 
     # TODO implement a check_params method to check the params
     #  using the JSON schema file.
@@ -24,12 +28,12 @@ class BaseModel(ConfigObject):
         filename (Str): Indicates where to store the model,
         if filename is None, this method returns a bytes array with the model.
         """
-        pass
+        raise NotImplementedError
 
-    @staticmethod
-    def load(filename):
+    @abstractmethod
+    def load(self, filename):
         """
         Restores an instance of a model
         filename (Str): Indicates where the model was stored.
         """
-        pass
+        raise NotImplementedError