+++ DashAI/back/models/scikit_learn/random_forest_classifier.py
@@ -7,7 +7,7 @@ from DashAI.back.models.tabular_classification_model import TabularClassificatio
 
 
 class RandomForestClassifier(
-    SklearnLikeModel, TabularClassificationModel, _RandomForestClassifier
+    TabularClassificationModel, SklearnLikeModel, _RandomForestClassifier
 ):
     """ """